refactor: HelpContent 컴포넌트 추가
